refactor(bloat): extract risk detectors + integration-coverage verifier into dedicated modules
The cross_component gate (#218) pushed two already-oversize files further
past the 300-LOC limit. Relocate cohesive clusters into new modules,
re-exported verbatim from their original homes (no behavior change), and
ratchet both bloat baselines down:

- risk_detectors.py: the diff-driven detectors lifted out of
  classify_complexity.py (430->330), re-exported so the iterate contract,
  cc.* attribute access, and the detector tests keep resolving them.
- integration_coverage.py + git_helpers.py: the cross-component gate plus
  the shared git wrappers lifted out of iterate_checks.py (1244->1122),
  re-exported so run_all_checks + the drift-sync test stay intact.

A neutral git_helpers module (not common.py, which would ratchet up; not a
3rd duplicate copy) lets the new verifier sub-module and iterate_checks
share one git-wrapper copy with no circular import.

All prior import paths preserved; full suite green (iterate plugin 396,
shared 3401, contract 21/21); ruff clean; anti-ratchet exit 0.

Run-ID: iterate-2026-06-13-risk-detector-extract
