--- .shipwright/agent_docs/iterates/iterate-2026-06-07-sbom-not-installed-vs-undeclared.json
@@ -1,10 +1,0 @@
-{
-  "run_id": "iterate-2026-06-07-sbom-not-installed-vs-undeclared",
-  "date": "2026-06-07T18:40:43.460001Z",
-  "type": "change",
-  "complexity": "medium",
-  "branch": "iterate/sbom-not-installed-vs-undeclared",
-  "tests_passed": true,
-  "spec": ".shipwright/planning/iterate/2026-06-07-sbom-not-installed-vs-undeclared.md",
-  "adr": "iterate-2026-06-07-sbom-not-installed-vs-undeclared"
-}
