+++ .shipwright/agent_docs/session_handoff.md
@@ -1,37 +1,35 @@
 ---
 canon_generated: true
-run_id: "iterate-2026-06-10-phase-hook-lifecycle"
+run_id: "iterate-2026-06-13-risk-detector-extract"
 phase: "iterate"
-reason: "iterate: phase-hook lifecycle + event-log integrity (WP1)"
-timestamp: "2026-06-12T22:41:27.823708+00:00"
+reason: "iterate: extract risk detectors + integration-coverage verifier (bloat ratchet)"
+timestamp: "2026-06-13T04:34:38.831317+00:00"
 ---
 
 # Session Handoff
 
-> Auto-generated 2026-06-12 22:41:27 UTC
+> Auto-generated 2026-06-13 04:34:38 UTC
 
 ## Session Info
 
-- **Session ID**: 30d44a6a-9835-4347-a9f2-b6b6a8b528f5
-- **Timestamp**: 2026-06-12 22:41:27 UTC
-- **Reason**: iterate: phase-hook lifecycle + event-log integrity (WP1)
+- **Session ID**: 553f1ff3-e4ba-4ee6-9e55-19f635416972
+- **Timestamp**: 2026-06-13 04:34:38 UTC
+- **Reason**: iterate: extract risk detectors + integration-coverage verifier (bloat ratchet)
 
 ## Last Iterate
 
-- **Run ID**: iterate-2026-06-12-reducibility-gate
-- **Date**: 2026-06-12T21:07:05.540726Z
+- **Run ID**: iterate-2026-06-10-phase-hook-lifecycle
+- **Date**: 2026-06-12T22:41:57.216562Z
 - **Type**: change
 - **Complexity**: medium
-- **Branch**: iterate/reducibility-gate
-- **ADR**: iterate-2026-06-12-reducibility-gate
+- **Branch**: iterate/2026-06-10-phase-hook-lifecycle
+- **ADR**: iterate-2026-06-10-phase-hook-lifecycle
 - **Tests passed**: True
-- **Spec**: .shipwright/planning/iterate/20260612-reducibility-gate.md
+- **Spec**: .shipwright/planning/iterate/2026-06-10-phase-hook-lifecycle.md
 
 ## Current Iterate Progress
 
-- **Branch**: iterate/2026-06-10-phase-hook-lifecycle
-- **Spec**: .shipwright/planning/iterate/2026-06-10-phase-hook-lifecycle.md
-- **Complexity**: medium (locked)
+- **Branch**: iterate/risk-detector-extract
 - **External Review Marker**: missing
 
 ### Mandatory replay on Resume
@@ -50,8 +48,8 @@ Before dispatching to the handoff's Remaining phase, run these if missing:
 
 ## Git State
 
-- **Branch**: iterate/2026-06-10-phase-hook-lifecycle
-- **Last Commit**: 60481718 chore(release): v0.25.0 (#223)
+- **Branch**: iterate/risk-detector-extract
+- **Last Commit**: a683d75b fix(pipeline): phase hooks resolve identity from stdin payload; atomic event dedup; failure event types (WP1) (#224)
 - **Uncommitted Changes**: Yes
 
 ## Config Files to Read
@@ -67,17 +65,17 @@ Before dispatching to the handoff's Remaining phase, run these if missing:
 
 | Event | Type | Source | Date |
 |-------|------|--------|------|
+| evt-e7fde4fc | work_completed | iterate (extract diff-driven risk detectors + integration-coverage verifier into dedicated modules to ratchet two bloat baselines down) | 2026-06-13 |
 | evt-8b8ef149 | work_completed | iterate (WP1: phase-session hooks resolve identity from the stdin payload (F1); atomic event-log dedup (F14); phase_failed/stale_stop_rejected event types (F15)) | 2026-06-12 |
 | evt-8c8f2132 | work_completed | iterate (Coerce explicit-null affected_frs/new_frs (and tests/review) in WorkEvent.from_dict) | 2026-06-12 |
 | evt-1c8dc50c | work_completed | iterate (Relocate resolve_main_repo_root from lib/events_log.py to lib/repo_root.py with a lazy back-compat re-export; migrate net-zero consumers; keep the two grandfathered consumers (iterate_checks, group_f) on the re-export to avoid ratcheting bloat.) | 2026-06-12 |
 | evt-e36182b6 | work_completed | iterate (Intelligent bloat gate: LOC-as-router -> falsifiable reducibility reviewer (closed catalog D/A/X/C/S/M/P/T + guardrails G1-G6); shared SSoT catalog + per-language idiom-map + reviewer dimensions across 3 surfaces + drift-protection test.) | 2026-06-12 |
-| evt-29b841b9 | work_completed | iterate (W2 phase-quality check SKIPs on an unresolvable run_id (mirror S2/S3); fixes the audit-context false-FAIL/false-PASS when no iterate run resolves; also fixes a latent empty-run_id crash) | 2026-06-12 |
 
 ## Recovery
 
 - **Pipeline**: 1 phases completed
-- **Total work events**: 168
-- **Last iterate**: change — WP1: phase-session hooks resolve identity from the stdin payload (F1); atomic event-log dedup (F14); phase_failed/stale_stop_rejected event types (F15) (2026-06-12)
+- **Total work events**: 169
+- **Last iterate**: change — extract diff-driven risk detectors + integration-coverage verifier into dedicated modules to ratchet two bloat baselines down (2026-06-13)
 - **Resume**: `/shipwright-iterate` for next change, or `/shipwright-run` for new pipeline
 
 ## Recent Decisions
