+++ .shipwright/agent_docs/triage_inbox.md
@@ -1,6 +1,6 @@
 # Triage Inbox
 
-> Auto-generated 2026-06-12T22:41:27.823708Z. Items waiting for triage decision.
+> Auto-generated 2026-06-13T04:34:38.831317Z. Items waiting for triage decision.
 > Promote via WebUI Triage tab (when v1b lands) or `shared/scripts/tools/triage_promote.py --id <id> --task-ref EXT:<ref>`.
 
 ## Status summary
