+++ .shipwright/compliance/sbom.md
@@ -1,6 +1,6 @@
 # Software Bill of Materials (SBOM)
 
-Generated: 2026-06-12T22:41:27.823708+00:00
+Generated: 2026-06-13T04:34:38.831317+00:00
 
 ## Summary
 
@@ -16,7 +16,7 @@ Generated: 2026-06-12T22:41:27.823708+00:00
 
 ```mermaid
 pie title License Distribution
-    "MIT" : 2
+    "MIT" : 4
     "Apache-2.0" : 1
 ```
 
@@ -35,8 +35,8 @@ pie title License Distribution
 
 | Package | Version | License |
 |---------|---------|---------|
-| pytest | 8.0.0 | - |
-| pytest-mock | 3.12.0 | - |
+| pytest | 8.0.0 | MIT |
+| pytest-mock | 3.12.0 | MIT |
 
 ## License Compliance
