+++ shipwright_events.jsonl
@@ -171,3 +171,4 @@
 {"intent":"change","spec_impact":"none","change_type":"tooling","none_reason":"Internal shared-lib relocation: resolve_main_repo_root moved to lib/repo_root.py plus a lazy back-compat re-export in events_log; consumer imports updated. Contract and all consumer behavior unchanged; no FR or product spec changed.","description":"Relocate resolve_main_repo_root from lib/events_log.py to lib/repo_root.py with a lazy back-compat re-export; migrate net-zero consumers; keep the two grandfathered consumers (iterate_checks, group_f) on the re-export to avoid ratcheting bloat.","v":1,"id":"evt-1c8dc50c","ts":"2026-06-12T20:49:23.721279+00:00","type":"work_completed","source":"iterate","commit":"","adr_id":"iterate-2026-06-12-repo-root-resolver-relocate","session":"bcb718a2-6a5f-41b1-ba60-3122da90f99b"}
 {"intent":"bug","description":"Coerce explicit-null affected_frs/new_frs (and tests/review) in WorkEvent.from_dict","spec_impact":"none","spec_impact_justification":"Bug fix in compliance deserialization; no functional requirement added, modified, or removed.","change_type":"compliance","none_reason":"Hardens compliance WorkEvent.from_dict against explicit-null events; no FR touched.","tests":{"passed":697,"total":697,"e2e_run":false},"v":1,"id":"evt-8c8f2132","ts":"2026-06-12T20:22:50.182497+00:00","type":"work_completed","source":"iterate","commit":"","adr_id":"iterate-2026-06-12-workevent-null-frs-coerce","session":"5194116e-24a8-4fd8-95b9-06465ff26727"}
 {"intent":"change","description":"WP1: phase-session hooks resolve identity from the stdin payload (F1); atomic event-log dedup (F14); phase_failed/stale_stop_rejected event types (F15)","spec_impact":"none","spec_impact_justification":"Framework correctness fix to Claude-Code phase-hook identity resolution + the record_event event-log writer; touches no FR.","change_type":"tooling","none_reason":"Framework hook plumbing + event-log writer correctness fix (deep-audit WP1); no functional requirement added or changed.","campaign":"2026-06-10-audit-2-manual","sub_iterate_id":"a2-1","tests":{"passed":3348,"total":3362,"e2e_run":false},"changed_files":["shared/scripts/lib/hook_session.py","shared/scripts/hooks/phase_session_start.py","shared/scripts/hooks/phase_session_stop.py","shared/scripts/hooks/phase_user_prompt_validate.py","shared/scripts/tools/record_event.py","shared/tests/test_hook_session.py","shared/tests/test_record_event_lifecycle_integrity.py","integration-tests/test_phase_hook_main_lifecycle.py","shipwright_bloat_baseline.json"],"v":1,"id":"evt-8b8ef149","ts":"2026-06-12T22:41:27.823708+00:00","type":"work_completed","source":"iterate","commit":"","adr_id":"iterate-2026-06-10-phase-hook-lifecycle","session":"30d44a6a-9835-4347-a9f2-b6b6a8b528f5"}
+{"intent":"change","description":"extract diff-driven risk detectors + integration-coverage verifier into dedicated modules to ratchet two bloat baselines down","spec_impact":"none","spec_impact_justification":"pure framework-internal refactor (verbatim relocation + re-export); no FR behavior change","change_type":"tooling","none_reason":"framework-internal module relocation + re-export; no functional requirement touched","changed_files":["plugins/shipwright-iterate/scripts/lib/classify_complexity.py","plugins/shipwright-iterate/scripts/lib/risk_detectors.py","shared/scripts/tools/verifiers/iterate_checks.py","shared/scripts/tools/verifiers/integration_coverage.py","shared/scripts/tools/verifiers/git_helpers.py","shipwright_bloat_baseline.json"],"tests":{"passed":3818,"total":3830,"e2e_run":false},"v":1,"id":"evt-e7fde4fc","ts":"2026-06-13T04:34:38.831317+00:00","type":"work_completed","source":"iterate","commit":"","adr_id":"iterate-2026-06-13-risk-detector-extract","session":"553f1ff3-e4ba-4ee6-9e55-19f635416972"}
